feat: page composition on placing-order page
